Made changes in response to inclusion of TCP/TLS as a possible network abstraction. Removed timestamp field from ProtocolDatagram structure. Removed assumption that ProtocolDatagram will be used on a UDP network. Removed assumption that ack timeout will always be enabled. Handled in DatagramChopper skipping of session layer options which can interfere with bulk sending.

+++ csharp/PortableIPC.Core/Abstractions/ISessionHandler.cs
@@ -34,9 +34,9 @@ namespace PortableIPC.Core.Abstractions
         //    last received id is larger than 0.
         // By so doing receiver can be conservative, and sender can have 
         // freedom in varying trend of window ids.
-        long NextWindowIdToSend { get; set; }
-        long LastWindowIdSent { get; set; }
-        long LastWindowIdReceived { get; set; }
+        int NextWindowIdToSend { get; set; }
+        int LastWindowIdSent { get; set; }
+        int LastWindowIdReceived { get; set; }
         int LastMaxSeqReceived { get; set; }
 
         // timeout api assumes only 1 timeout can be outstanding at any time.
